tests

+++ tests/tests_cloud/utils.py
@@ -5,6 +5,5 @@ from tests_cloud import LIGHTNING_TEST_STORAGE_DIR
 
 
 def cleanup():
-    if os.getenv("LIGHTNING_MODEL_STORE_TESTING"):
-        if os.path.isdir(LIGHTNING_TEST_STORAGE_DIR):
-            shutil.rmtree(LIGHTNING_TEST_STORAGE_DIR)
+    if os.getenv("LIGHTNING_MODEL_STORE_TESTING") and os.path.isdir(LIGHTNING_TEST_STORAGE_DIR):
+        shutil.rmtree(LIGHTNING_TEST_STORAGE_DIR)
